Solved offset problem in seek for zip, gzio and tar objects

+++ packages/misc/gzio/c_src/gzfile_stream.h
@@ -53,7 +53,7 @@ namespace GZIO {
     virtual bool privateEof() const;    
     virtual size_t privateWrite(const char *buf, size_t size);
     virtual size_t privateRead(char *buf, size_t max_size);
-    virtual off_t privateSeek(int whence, int offset);
+    virtual off_t privateSeek(int whence, long offset);
     
   private:
